Crate holochain_core

The library implementing the holochain pattern of validation rules + local source chain + DHT

Modules

action

agent

consistency

context

dht
DHT is the module that handles the agent's local shard of data and p2p communications

entry
This module extends Entry and EntryType with the CanPublish trait.

instance

logger
This logger is the logger that's attached to each Holochain application which is separate from standard logging via the log crate warn! info! debug! logging that gets emitted globaly from the conductor.

network

nucleus

persister

scheduled_jobs

signal

state

workflows

Macros

link_zome_api