Module holochain_core::context

Structs

Context
Context holds the components that parts of a Holochain instance need in order to operate. This includes components that are injected from the outside like logger and persister but also the store of the instance that gets injected before passing on the context to inner components/reducers.

Functions

get_dna_and_agent

test_memory_network_config
Create an in-memory network config with the provided name, otherwise create a unique name and thus network using snowflake. This is the base function that many other text_context* functions use, and hence they also require an optional network name. The reasoning for this is that tests which only require a single instance may simply pass None and get a unique network name, but tests which require two instances to be on the same network need to ensure both contexts use the same network name.