Module holochain_core::nucleus::ribosome::api

Module for ZomeApiFunctions ZomeApiFunctions are the functions provided by the ribosome that are callable by Zomes.

Modules

call

capabilities

commit

crypto

debug

emit_signal

entry_address

get_entry

get_links

get_links_count

init_globals

keystore

link_entries

query

remove_entry

remove_link

send

sign

sleep

update_entry

verify_signature

Enums

ZomeApiFunction
Enumeration of all the Zome Functions known and usable in Zomes. Enumeration can convert to str.

Type Definitions

ZomeApiResult