Enum holochain_core::nucleus::ribosome::api::ZomeApiFunction

#[repr(usize)]

pub enum ZomeApiFunction {
    MissingNo,
    Abort,
    Debug,
    CommitAppEntry,
    GetAppEntry,
    UpdateEntry,
    RemoveEntry,
    InitGlobals,
    Call,
    LinkEntries,
    GetLinks,
    Query,
    EntryAddress,
    Send,
    Sleep,
    RemoveLink,
}

Enumeration of all the Zome Functions known and usable in Zomes. Enumeration can convert to str.

Variants
MissingNo

Error index for unimplemented functions

Abort

Abort is a way to receive useful debug info from assemblyscript memory allocators message: mem address in the wasm memory for an error message filename: mem address in the wasm memory for a filename line: line number column: column number

Debug

Zome API send debug information to the log debug(s: String)

CommitAppEntry

Commit an app entry to source chain commit_entry(entry_type: String, entry_value: String) -> Address

GetAppEntry

Get an app entry from source chain by key (header hash) get_entry(address: Address) -> Entry

UpdateEntryRemoveEntryInitGlobals

Init Zome API Globals hc_init_globals() -> InitGlobalsOutput

Call

Call a zome function in a different capability or zome hc_call(zome_name: String, cap_token: Address, fn_name: String, args: String);

LinkEntriesGetLinksQueryEntryAddress

Pass an entry to retrieve its address the address algorithm is specific to the entry, typically sha256 but can differ entry_address(entry: Entry) -> Address

SendSleepRemoveLink
Methods

impl ZomeApiFunction

pub fn as_fn(&self) -> fn(_: &mut Runtime, _: &RuntimeArgs) -> ZomeApiResult