Function holochain_core::nucleus::ribosome::api::query::invoke_query

pub fn invoke_query(runtime: &mut Runtime, args: &RuntimeArgs) -> ZomeApiResult

ZomeApiFunction::query function code args: [0] encoded MemoryAllocation as u64 Expected complex argument: ? Returns an HcApiReturnCode as I64

Specify 0 or more simple or "glob" patterns matching EntryType names, returning Vec.

The empty String or an empty Vec matches all. The '*' glob pattern matches all simple EntryType names (with no '/'), while the ** pattern matches everything (use "" or [] for efficiency).

[] [""] ["**"]

Namespaces (groups of related EntryType names) can be queried easily, eg:

["name/*"]

Several simple names and/or "glob" patterns can be supplied, and are efficiently searched for in a single pass using a single efficient Regular Expression engine:

["name/*", "and_another", "something_else"]

EntryType names can be excluded, eg. to return every simple (non-namespaced) EntryType except System:

["[!%]*"]

To match a pattern, including all namespaced EntryType names, eg. every EntryType except System:

["**/[!%]*"]

The following standard "glob" patterns are supported:

Pattern Match ======= ===== . One character (other than a '/') [abcd] One of a set of characters [a-d] One of a range of characters [!a-d] Not one of range of characters {abc,123} One of a number of sequences of characters * Zero or more of any character **/ Zero or more of any namespace component