Struct holochain_core::state::StateWrapper

pub struct StateWrapper { /* fields omitted */ }

This type wraps (decorates) InnerState with an option and re-exports and delegates all methods of InnerState. It owns the InnerState and keeps it in a Option so that it can be dropped explicitly. It also adds a function drop(&mut self) which sets the option to None which will drop the InnerState.

Methods

impl StateWrapper

pub fn drop_inner_state(&mut self)

pub fn new(context: Arc<Context>) -> Self

pub fn new_with_agent(context: Arc<Context>, agent_state: AgentState) -> Self

pub fn new_with_agent_and_nucleus(
context: Arc<Context>,
agent_state: AgentState,
nucleus_state: NucleusState
) -> Self

pub fn reduce(&self, action_wrapper: ActionWrapper) -> Self

pub fn nucleus(&self) -> Arc<NucleusState>

pub fn agent(&self) -> Arc<AgentState>

pub fn dht(&self) -> Arc<DhtStore>

pub fn network(&self) -> Arc<NetworkState>

pub fn get_headers(
&self,
entry_address: Address
) -> Result<Vec<ChainHeader>, HolochainError>

pub fn conductor_api(&self) -> ConductorApi

pub fn history(&self) -> HashSet<ActionWrapper>